requirements.txt
torch==2.2.0
torch==2.2.0
langchain-community==0.0.17
python-box==7.1.1
langchain==0.1.5
fastapi==0.109.0
uvicorn==0.27.0
sentence-transformers==2.3.1
chromadb==0.4.22
ctransformers==0.2.27
langchain_google_genai
google.generativeai